Move changes of adding plugin meta to PluginSuppression.php service
