+++ src/PluginSuppression.php
@@ -18,6 +18,7 @@ use AmpProject\AmpWP\Infrastructure\Registerable;
 use AmpProject\AmpWP\Infrastructure\Service;
 use WP_Block_Type_Registry;
 use WP_Hook;
+use WP_Screen;
 use WP_Term;
 use WP_Widget;
 use WP_User;
@@ -83,6 +84,7 @@ final class PluginSuppression implements Service, Registerable {
 	public function register() {
 		add_filter( 'amp_default_options', [ $this, 'filter_default_options' ] );
 		add_filter( 'amp_options_updating', [ $this, 'sanitize_options' ], 10, 2 );
+		add_filter( 'plugin_row_meta', [ $this, 'filter_plugin_row_meta' ], 10, 3 );
 
 		add_filter(
 			'register_block_type_args',
@@ -490,5 +492,41 @@ final class PluginSuppression implements Service, Registerable {
 			'plugin' === $source['type'] &&
 			in_array( $source['name'], $suppressed_plugins, true )
 		);
+	}
+
+	/**
+	 * Add meta if plugin is suppressed in AMP page.
+	 *
+	 * @param array  $plugin_meta An array of the plugin's metadata.
+	 * @param string $plugin_file Path to the plugin file relative to the plugins directory.
+	 * @param array  $plugin_data An array of plugin data.
+	 *
+	 * @return array An array of the plugin's metadata
+	 */
+	public function filter_plugin_row_meta( $plugin_meta, $plugin_file, $plugin_data ) {
+
+		$screen = get_current_screen();
+
+		if ( ! $screen instanceof WP_Screen || 'plugins' !== $screen->id ) {
+			return $plugin_meta;
+		}
+
+		$suppressed_plugins = AMP_Options_Manager::get_option( 'suppressed_plugins' );
+		$plugin_slug        = isset( $plugin_data['slug'] ) ? $plugin_data['slug'] : false;
+
+		if ( empty( $plugin_slug ) ) {
+			$plugin_slug = $this->plugin_registry->get_plugin_slug_from_file( $plugin_file );
+		}
+
+		if ( isset( $suppressed_plugins[ $plugin_slug ] ) ) {
+			$plugin_meta[] = sprintf(
+				'<a href="%s" aria-label="%s">%s</a>',
+				esc_url( admin_url( 'admin.php?page=amp-options' ) ),
+				esc_attr__( 'Visit AMP Settings', 'amp' ),
+				esc_html__( 'Suppressed on AMP Pages', 'amp' )
+			);
+		}
+
+		return $plugin_meta;
 	}
 }
